feat: Add static analysis, tests, pre-commit hooks, and CI
- Configure Ruff linting/formatting for backend (pyproject.toml)
- Configure Prettier for frontend (.prettierrc, format scripts)
- Strengthen mobile analysis_options (strict-casts, prefer_const)
- Add pre-commit hooks (ruff, prettier, flutter-analyze, trailing-ws)
- Add 36 backend tests (models, business logic, endpoints)
- Add 36 frontend tests (utils, hooks, components via Vitest)
- Add 28 mobile tests (formatters, position_utils, fixture model/repo)
- Add GitHub Actions CI with 3 parallel jobs
- Add Claude Code auto-format hooks and CLAUDE.md commands reference

+++ mobile/lib/features/home/screens/home_screen.dart
@@ -164,12 +164,12 @@ class _HomeScreenState extends ConsumerState<HomeScreen> {
 
   void _runAnalysis(int teamId, dynamic data) {
     final bank = data.history?.isNotEmpty == true
-        ? (data.history!.last.bank / 10)
+        ? (data.history!.last.bank / 10) as double
         : 0.0;
     ref.read(analysisStateProvider.notifier).analyze(
           teamId: teamId,
           moneyInBank: bank,
-          freeTransfers: data.freeTransfers ?? 1,
+          freeTransfers: (data.freeTransfers ?? 1) as int,
           authToken: ref.read(savedAuthTokenProvider),
         );
   }
